requirements.txt
flask
pandas
requests
mysql-connector-python